Refactor `getDiscountGroups` to return an array of `DiscountGroupDTO` objects, remove `DiscountGroupsDTO`, and update type handling for `discountPercent`.

+++ src/DiscountGroup/DTO/DiscountGroupDTO.php
@@ -10,7 +10,7 @@ final class DiscountGroupDTO extends DTO
 {
     public int $_cloudId;
     public bool $deleted;
-    public int $discountPercent;
+    public float $discountPercent;
     public bool $display;
     public ?string $externalId = null;
     public int $id;
